requirements.txt
requests==2.32.3
pyspark==3.5.4
pyyaml

